requirements.txt
asgiref==3.7.2
Django==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-phonenumber-field==7.3.0
django-simple-history==3.8.0
djangorestframework==3.14.0
gunicorn==23.0.0
phonenumbers==8.13.27
pillow==11.1.0  
psycopg2-binary==2.9.10  
python-dotenv==1.0.1 
pytz==2023.3.post1
sqlparse==0.4.4
bleach
django_bower
requests
certifi
stripe
